requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-WTF==1.2.1
Flask-Login==0.6.3
Flask-Mail==0.9.1
Werkzeug==3.0.1
WTForms==3.1.1
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
bcrypt==4.1.2
pyotp==2.9.0
qrcode==7.4.2
Pillow==10.1.0
cryptography==41.0.8
bleach==6.1.0
email-validator==2.1.0
requests==2.31.0
pandas==2.1.4
plotly==5.17.0
folium==0.15.1
gunicorn==21.2.0
python-dateutil==2.8.2
